feat: Update user routes to use resourceful routing
This commit updates the user routes in the `web.php` file to use resourceful routing. The previous individual route declarations for index, create, store, show, edit, update, and destroy are commented out, and a single `Route::resource` declaration is added to handle all the CRUD operations for the `UserController` class. This change simplifies the code and follows the convention for RESTful APIs.